Fix downloading extra data in libflatpak
The download was getting stuck when launched from within a thread as is
the case when using GNOME Software.

This patch simplifies the code that downloads the extra data and drops
the use of a mainloop, calling the libsoup related code synchronously
instead. It also now returns an error when an issue occurred while the
data was being downloaded.